Finished!

+++ src/main/java/com/example/elastic_search_project/repository/CustomerRepository.java
@@ -1,12 +1,10 @@
 package com.example.elastic_search_project.repository;
 
+import java.util.List;
+
 import com.example.elastic_search_project.document.Customer;
 import org.springframework.data.elasticsearch.repository.ElasticsearchRepository;
 
-import java.util.List;
-
-public interface CustomerRepository extends ElasticsearchRepository<Customer, Long> {
-
+public interface CustomerRepository extends ElasticsearchRepository<Customer, Integer> {
     List<Customer> findByFullName(String fullName);
-
 }
